requirements.txt
# Frontend
streamlit==1.45.1
pandas==2.2.3
python-dotenv==1.1.0

# Backend (para futura implementação)
flask==3.0.0
werkzeug==3.0.1

# Utilitários comuns
numpy==2.2.6